meme-a-tron/backend/requirements.txt
fastapi
uvicorn
pillow
openai